Fix the ELF section NOBITS fix again.
Obviously, we can't assume that the section numbers match.  For instance,
the ELF file containing debuginfo probably won't have its .eh_frame section;
that will probably be in the actual stripped ELF binary/lib file.  However,
when we go searching for it in the ELF file, we have to rescan to get the
right section index...