Actually finish decl/defn resolution!
This was all-but-done, and non-live testing didn't reveal it.  I even
had an XXX to do something about it.  Basically, global debugfile load
post passes like declaration/definition resolution have to give the
backend a chance to update its internal symbol pointers (i.e., symbols
that pointed to the declaration that is being replaced must be updated
to point to the definition).  So that's done, easy.