Add thread contexts for Xen and Xen-Process targets.
Basically, the Xen target detects whether the current thread is in kernel or
user space.  Since this is a non-SMP target, we can make the assumption that
any non-current thread is in the kernel; sometimes, the current thread might
be in userspace.  With this we can think of supporting reads or writes to
either the kernel or user register context.

Right now, this is used to allow a Xen Process target to read only user-level
registers; it still needs to be fleshed out.