A better way (?) to deal with DWARF C++-oriented AT_specification.
We had basically attempted to just move the definition symbol into the
scope of the declaration, and (optionally) remove the declaration symbol
(on the theory that it is typically useless once we have harvested its
attributes).

Sadly, this was short-sighted.  Declaration DIEs that are later referenced
via specification by definition DIEs sometimes *do* contain member symbols,
and sometimes they are not duplicated in the definition symbol.

We had to be smarter about handling members of declarations that did not
reappear in definitions as member defn symbols linked back to the member
decls.  But, we can't actually do this until we have parsed all the members.
So it gets pretty hard.

This is not implemented for the PARTIALSYM nor type compression features
yet.  That is *really* hard.