No Probes Left Behind!
Forcibly remove probepoints in the "ext" case.

When the target lib needs to remove a probepoint asynchronously w.r.t. to
the target's flow of execution, we want to force that to happen.  Forcing
will remove the probepoint even if it is being handled... which sometimes
we want to do (like on signals).

No Probes Left Behind!