Add first part of multithread support in Xen Process.
This is definitely the quickie approach.  Basically, in the Xen target
(the underlying target), if a user thread hits a breakpoint and we do
not recognize it as attached to an overlay, see if it is in a thread
group for a process that *is* attached to an overlay.  If so, we force
the overlay to attach to it -- well, we expect it that it will -- and
to handle the exception.

There's a simple model in the target library supporting this; overlay
targets can use thread_attach_overlay() to attach the overlay to an
underlying thread.  The Xen target now basically forces one overlay
per thread group, not per thread... that is a valid design decision.
Other backends might do differently.  It's flexible.

Also moved the code that detects if a thread being used to instantiate
the overlay is or is not the thread group leader from Xen-Process to
Xen target where it should be; the Xen target will only instantiate
overlay targets atop thread group leaders.

This is the first iteration of this new feature; the Xen-Process target
does not yet try to do anything about threads exiting, nor does it
actively detect new ones itself.