CFI: handle addrs not resolveable to a symbol by instrumenting code blocks.
When control transfers to a new address that we cannot find a symbol for,
instead of throwing up our hands, check if the address is in a code block
that is safe to disasm (meaning, the new address is contained in a code
block pointed to by a valid symbol -- so that we can safely do x86 variable-
length instruction disassembly starting at a known, valid address).

Then, instrument control flow instructions in that block instead of the
symbol.

cfi_check has been improved to take raw addresses that are instrumented
in this fashion as part of the root set, too.

There is a new probe_register_block_instrs that does basically the same
thing probe_register_function_instrs does, but operates on a (target,
startaddr,endaddr) tuple instead of a function symbol.