Handle a likely race condition in xc_domain_getinfo().
Now that we have multiple domain probing support, we're seeing times where
xc_domain_getinfo() returns without error, but also doesn't set any of the
domain status flags (dying,crashed,shutdown,paused,blocked,running).  If we
get into this situation, we can't do much more than wet our pants.  We could
assume the domain is paused, but that is probably not true.  Based on my
experience, the domain is hosed at this point.

Anyway, this commit introduces an alternate strategy: poll!  In practice, a
single retry works, it seems.  Right now, polling is implemented by a select()
loop with an interval of 10us.  Probably way too much, and probably I should
go to pselect to get nanosecond level resolution.  But had to try it first,
and now I have to commit it!

I'm not sure what the real race problem is.  One (really, really) unlikely
scenario could be as follows.  If the xc call does not get the updated
status (paused) before the hypervisor reports the pause event on the
evtchn.  The thing that doesn't maek sense is that this would be a very
slow race.  Consider: the VIRQ demux daemon running in dom0 gets
notified via evtchn; the dom0 kernel ctx switches to the demux clients
so they can run dominfo() and check their respective domain's status.
That status check usually reports correct, but not always.  But there's a
*ton* of time for dominfo() to have gotten the right info into its memory
structures, unless there's some kind of lazy update happening here.  Weird!
It's too slow to be believable.

I could guess about other possible reasons... but pointless.