Change psaction's "stop" command into "sig".
This sends an arbitrary signal to a process.  The stop command
requires a signal name or number, and an rfilter that determines
which processes the signal will be "sent" to.  For instance,

  $ psaction -m domain sig quit R::A:ping

will send a SIGQUIT to any processes whose task_struct->comm matches
the regexp "ping".

Right now, STOP/CONT/TSTP/TTIN/TTOU do not work; all others work.