Fix the "sig" command to work for SIGSTOP.
Just don't set the SIGNAL_GROUP_EXIT flag.  I think I did this for the
kill command to force the signal to get broadcast to the whole process
groups, but I can't be sure.  In any event, it isn't necessary to set it
to achieve a kill of a single-threaded process.  And realistically, if
you kill the primary thread of a multithreaded process, it's going to go
away... so this kinda just works.  It's only if you tried to kill a
secondary thread of a process group via the sig command that the entire
group might not killed.  But I'm still speculating there.

Also -- this does not, and never will work, for SIGCONT.  Meaning that
if the process is stopped, and you try to continue it via psaction, that
won't work -- because this requires a whole bunch of kernel data structure
manipulation.  In particular, we must reach deep into the scheduler and
adjust the runqueues... choose a processor for the stopped thread to run
on... all kinds of crap.  Also, even a SIGKILL on a stopped process won't
work, for basically the same reason.  We just can't wake a stopped process
up.  Well, we can for non-SMP systems, or if we make all kinds of assumptions
that do an end-around the scheduler, and if we assume we can just blast in
and mess with (potentially) locked runqueues.

For the 2.6.18 kernel, all this magic is in sched.c:try_to_wake_up , and
in the stuff it calls.