"Improve" the autoparsing code yet again.
This time, change the format to a tagged result (i.e.,
RESULT:: %format ::RESULT).  This of course is just a gigantic
mess.  It would be better obviously if tools could generate XML
or b64encode their data, but we chose to not do it that way, so
we're stuck with this crappy hacky autoparsing.  Argh.  Anyway,
have start/stop tags of course improves parseability, which was
becoming impossible.  There are still potential for problems this
way, but it is reduced further.