Handle the inferred single step case, but now per-thread.
We've long known about the case where we command a single step but
by the time we get the debug exception the debug status register is
cleared (so we can't really tell what happened).  This could be a Xen
problem; a side effect of reusing the debug registers the guest kernel
(might) be using; or something else.  Anyway, it happens.  So in this
case, we infer a single step.  For some reason, I had only chosen to
match a single step at the probepoint in the last stepped thread...
but since the guest kernel might well switch threads prior to stepping
the first thread, and in the meantime step another thread -- we have
to check which thread is currently executing and match that thread's
probepoint.

I'm not sure why I didn't do this before... it's completely obvious.
Makes me wonder if this is going to introduce a bug.  But it's not
really solveable; if the debug registers don't have the expected
content, we have to make a guess.