Make SPF able to auto-backtrace overlay targets in various ways.
This allows the user to specify that they want either the report()
or bt() functions to print backtraces of threads all the way up
the overlay target stack.  These overlay targets are generated as
needed and destroyed immediately.  report() allows you to print all
thread info + backtrace for any of the threads and overlays being
reported on; bt() is simpler and only backtraces up the overlay
target stack.  See the docs for the details...