Improve CFA-based unwinding.
Previously, I had made the assumption that there would be a CFA FDE
associated with the start address of the symbol associated with the
current frame -- to lookup the FDE to get to the previous frame.  This
assumption allowed me to index the FDE info in a simple hashtable
rather than in one of my cl* Judy array wrappers, which allow range-
based searching.  Moreover, when I first did this, I didn't have support
for the clrangesimple_t structure, a better match for this case.  So,
not sure why I used hashtables instead of clrange before, but now
I've undone the assumption using clrangesimple_t.

This means we can always, trivially, find the FDE associated with the
current frame's IP address; this shouldn't be much more costly than
a hashtable either.  clrange_t would have been more costly.