Improve debuginfo_lookup_addr.
If we only were able to find a root symbol from debuginfo (a root
symbol corresponds to a source file, which is almost never useful),
try to find a better symbol from the binfile, and return it if we
find one.

Not sure if this really what we want?  The binfile symbol isn't going
to have much metadata associated with it... we should figure a
way to expose this to the user, perhaps.