Add a more powerful address->symbol lookup function.
Sometimes we want debuginfo symbols; sometimes we want binfile symbols.
In this particular case, when printing backtraces, we need the debugfile
symbol, but sometimes that symbol ends up being the CU-file-wide symbol --
and the binfile symbol might be a more precise name for the function in
the stack frame (i.e., if the function is an assembly label, the nearest
containing debugfile symbol will be the assembly source file symbol (the
CU symbol); but there might well be an ELF symbol for that address, or
one "containing" it.  That is the better symbol to use in the backtrace.

It is unfortunate (in some ways) that we care about the distinction
between debugfile and binfile symbols, but CFA currently depends
on debugfile symbols.  This is an implementation artifact, not a strict
dependency... it exists because of the way debugfile->ops is required
to do CFA.  This should be fixed, sigh.