If elfutils calls back for a section we think is OOB, pretend it is unmapped.
Probably this is happening because when we load debuginfo from a pointed-to
binfile, we create the binfile_instance based on the pointing-to binfile --
and that one is probably stripped -- whereas the pointed-to binfile is
probably not stripped and thus elfutils wants to know (for relocatable
binfiles) where the section(s) got mapped.

So, satisfy elfutils by telling it that the section is just unmapped; and
hope for the best.

Yes, this "fix" assumes that section index numbers are consistent across
stripped and unstripped binfiles.  Not good... we should really generate
new binfile instances for the pointed-to binfiles!