Fix (?) up %sp load off kernel stack.
I have no idea why this was commented out; hopefully just lint that
got committed by mistake.  The saved %sp and %ss might not be valid
if the kernel thread was interrupted in the kernel, but that's no
reason not to set it anyway, since we don't check for that case...
hm.