Bug: fixup load of kernel context regs for userspace threads, etc.
If the thread we're loading off the stack was a userspace thread, its
userspace regs are on the stack as ptregs; and then the best we can do
for the kernel regs are to load them out of the thread struct (i.e.,
ip, sp).  Also, do that if we didn't have enough type info from
debuginfo -- i.e., no ptregs info -- best effort.