Handle a special case where a PV guest is in the Xen hypercall page.
In this case, we can't load the current thread, because (at least
sometimes!) the current thread_info->task pointer is messed up --
like 0xffffffff00000000 .  This is strange, but observable.  So
instead of failing, we just load init_task as the task.  This seems
to be the right thing to do... we're on that stack in the hypercall.
At least it allows us to backtrace out of the hypercall stack.

However, the OS Linux personality has to read the hypercall_page
symbol to find it; and then the personality sets a special config
key that the target driver can read (in this case, the Xen driver
saves it off, but it doesn't currently do anything with it).

I believe this should fix the behavior we used to see where the
current thread failed to load on Xen PVMs.  Details supporting
my analysis:

My IP was in xen_hypercall_sched_op on a 3.8.x PV.  The current
thread_info (pointed to by %gs+static_offset because it's percpu data on
x86_64, and percpu data is indexed off the %gs pointer) in this case is
the global thread_info associated with the init_task task struct (both
structs point to each other).  You can check that the default
thread_info static data is correctly, statically initialized by reading
the vmlinux binary.

But by the time I am in xen_hypercall_sched_op (on PV x86_64), the
default thread_info->task pointer is munged so it no longer points to
init_task, and instead points to 0xffffffff00000000, and that is not valid.
I took a fast look in the kernel src, especially the entry.S asm, and
can't see this manipulation -- but it's the kind of thing that is easy
to hide.  Maybe this is a "convenient" way of detecting illegal uses of
current() in interrupt context, or when making a hypercall (current()
is the current task, basically *(%gs+thread_info_offset)).

Now, I *did* have a guard against loading *current() in interrupt
context (it's only valid in process context) -- I check the
thread_info->preempt_count field to see if we're in an interrupt context
-- but when making a hypercall we are not in interrupt context,
apparently (or at least not necessarily in one -- and presumably
xen_hypercall_sched_op is a yield via the idle path).